refactor(server): extract relationship_queries.go
Move cluster relationship CRUD and cluster membership functions.

Part of #75.
